fixed admin bug for popup (#227)
* fixed admin bug for popup

* added fe change for admin removing their own acc

* guard not let curr user demote/promote/remove

+++ frontend/src/components/organisms/userCard/UserCard.jsx
@@ -25,7 +25,7 @@ export const UserCard = ({ user_id }) => {
   const makeUserAdmin = () => {
     console.log(currUser);
     fetcher({
-      url: `/admins/${currUser?.id}?user_id=${currUser.id}`,
+      url: `/admins/${currUser?.id}?user_id=${user.id}`,
       method: 'PUT',
       token: currUser.accessToken,
       body: {
@@ -43,13 +43,13 @@ export const UserCard = ({ user_id }) => {
       .catch((error) => {
         console.log(error);
         triggerUserFetch();
-        //showPopup(popupStates.ERROR, error.message); // TODO fix popping up for
+        showPopup(popupStates.ERROR, error.message); // TODO fix popping up for
       }); // ordinary users
   };
 
   const makeUserNotAdmin = () => {
     fetcher({
-      url: `/admins/${currUser?.id}?user_id=${currUser.id}`,
+      url: `/admins/${currUser?.id}?user_id=${user.id}`,
       method: 'PUT',
       token: currUser.accessToken,
       body: {
@@ -67,28 +67,33 @@ export const UserCard = ({ user_id }) => {
       .catch((error) => {
         console.log(error);
         triggerUserFetch();
-        //showPopup(popupStates.ERROR, error.message); // TODO fix popping up for
+        // showPopup(popupStates.ERROR, error.message); // TODO fix popping up for
       }); // ordinary users
   };
 
   const removeUser = () => {
-    console.log(currUser);
-    fetcher({
-      url: `/admins/${currUser?.id}/users/${currUser.id}`,
-      method: 'DELETE',
-      token: user.accessToken,
-    })
-      .then((res) => {
-        console.log('User deleted', res);
-        showPopup(popupStates.SUCCESS, 'User deleted!');
-        setTimeout(() => {
-          router.push('/admin');
-        }, 500);
+    
+    if(user_id == currUser.id){
+      showPopup(popupStates.WARNING, "Cannot remove your own account!")
+    } else {
+      console.log(currUser);
+      fetcher({
+        url: `/admins/${currUser?.id}/users/${user.id}`,
+        method: 'DELETE',
+        token: user.accessToken,
       })
-      .catch((error) => {
-        console.log(error);
-        // showPopup(popupStates.ERROR, error.message); // TODO fix popping up for ordinary
-      }); // users
+        .then((res) => {
+          console.log('User deleted', res);
+          showPopup(popupStates.SUCCESS, 'User deleted!');
+          setTimeout(() => {
+            router.push('/admin');
+          }, 500);
+        })
+        .catch((error) => {
+          console.log(error);
+          // showPopup(popupStates.ERROR, error.message); // TODO fix popping up for ordinary
+        }); // users
+    }
   };
 
   return (
@@ -109,29 +114,32 @@ export const UserCard = ({ user_id }) => {
           <p className="leading-6">Email address: {user?.email}</p>
         </div>
       </div>
-      <div className="flex flex-col justify-around min-h-[6em] grow">
-        <Button onClick={() => removeUser()} variant={'error'} className="">
-          Remove User
-        </Button>
-        {user && user.is_admin == 0 && (
-          <Button
-            onClick={() => {
-              makeUserAdmin();
-            }}
-          >
-            Make user Admin
+      {
+        user_id != currUser.id && 
+        <div className="flex flex-col justify-around min-h-[6em] grow">
+          <Button onClick={() => removeUser()} variant={'error'} className="">
+            Remove User
           </Button>
-        )}
-        {user && user.is_admin == 1 && (
-          <Button
-            onClick={() => {
-              makeUserNotAdmin();
-            }}
-          >
-            Demote Admin
-          </Button>
-        )}
-      </div>
+          {user && user.is_admin == 0 && (
+            <Button
+              onClick={() => {
+                makeUserAdmin();
+              }}
+            >
+              Make user Admin
+            </Button>
+          )}
+          {user && user.is_admin == 1 && (
+            <Button
+              onClick={() => {
+                makeUserNotAdmin();
+              }}
+            >
+              Demote Admin
+            </Button>
+          )}
+        </div>
+      }
     </div>
   );
