+++ frontend/src/pages/admin/index.jsx
@@ -7,10 +7,11 @@ import { useUsers } from '@/hooks/useUsers';
 import { useUserContext } from '@/contexts/UserContext';
 import { fetcher } from '@/lib/api';
 import { popupStates, usePopupContext } from '@/contexts/PopupContext';
+import { Pulser } from '@/components/atoms/pulser/Pulser';
 
 const Admin = () => {
   const router = useRouter();
-  const {data: users, error} = useUsers();
+  const {data: users, loading, error, triggerUsersFetch} = useUsers();
   const {user: user} = useUserContext();
   const { showPopup } = usePopupContext();
 
@@ -42,18 +43,24 @@ const Admin = () => {
   }, [users])
 
   const removeUserById = (id) => {
-    console.log(user)
-    fetcher({
-      url: `/admins/${user?.id}/users/${id}`,
-      method: 'DELETE',
-      token: user.accessToken,
-    }).then((res) => {
-      console.log('User deleted', res);
-      showPopup(popupStates.SUCCESS, 'User deleted!');
-    }).catch((error) => {
-          console.log(error);
-          // showPopup(popupStates.ERROR, error.message); 
-        });                                               
+    if(id == user.id){
+      showPopup(popupStates.WARNING, "Cannot remove your own account!")
+    } else {
+      console.log(user)
+      fetcher({
+        url: `/admins/${user?.id}/users/${id}`,
+        method: 'DELETE',
+        token: user.accessToken,
+      }).then((res) => {
+        console.log('User deleted', res);
+        showPopup(popupStates.SUCCESS, 'User deleted!');
+        triggerUsersFetch();
+      }).catch((error) => {
+            console.log(error);
+            // showPopup(popupStates.ERROR, error.message); 
+            triggerUsersFetch();
+          });        
+    }                                       
   };
 
   const handleRemove = (e, user_id) => {
@@ -65,6 +72,18 @@ const Admin = () => {
   const handleUserClick = (user_id) => {
     router.push('admin/viewuser/' + user_id);
   };
+
+  if (loading) {
+    return (
+      <main>
+        <section>
+          <Pulser />
+          <Pulser />
+          <Pulser />
+        </section>
+      </main>
+    );
+  }
   
 
   return (
